backend/requirements/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
gunicorn==20.1.0 
Django==4.2.2
django-filter==23.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==6.0.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.13.3
Pillow==10.0.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
